+++ src/filmgrain_tmpl_16.rs
@@ -426,7 +426,7 @@ unsafe extern "C" fn generate_grain_y_c(
     let bitdepth_min_8: libc::c_int = 32 as libc::c_int
         - clz(bitdepth_max as libc::c_uint) - 8;
     let mut seed: libc::c_uint = (*data).seed;
-    let shift: libc::c_int = 4 as libc::c_int - bitdepth_min_8
+    let shift: libc::c_int = 4 - bitdepth_min_8
         + (*data).grain_scale_shift;
     let grain_ctr: libc::c_int = (128 as libc::c_int) << bitdepth_min_8;
     let grain_min: libc::c_int = -grain_ctr;
@@ -499,7 +499,7 @@ unsafe extern "C" fn generate_grain_uv_c(
     let mut seed: libc::c_uint = (*data).seed
         ^ (if uv != 0 { 0x49d8 as libc::c_int } else { 0xb524 as libc::c_int })
             as libc::c_uint;
-    let shift: libc::c_int = 4 as libc::c_int - bitdepth_min_8
+    let shift: libc::c_int = 4 - bitdepth_min_8
         + (*data).grain_scale_shift;
     let grain_ctr: libc::c_int = (128 as libc::c_int) << bitdepth_min_8;
     let grain_min: libc::c_int = -grain_ctr;
@@ -652,15 +652,15 @@ unsafe extern "C" fn sample_lut(
 ) -> entry {
     let randval: libc::c_int = (*offsets.offset(bx as isize))[by as usize];
     let offx: libc::c_int = 3 as libc::c_int
-        + (2 as libc::c_int >> subx)
-            * (3 as libc::c_int + (randval >> 4));
+        + (2 >> subx)
+            * (3 + (randval >> 4));
     let offy: libc::c_int = 3 as libc::c_int
-        + (2 as libc::c_int >> suby)
-            * (3 as libc::c_int + (randval & 0xf as libc::c_int));
+        + (2 >> suby)
+            * (3 + (randval & 0xf as libc::c_int));
     return (*grain_lut
         .offset(
-            (offy + y + (32 as libc::c_int >> suby) * by) as isize,
-        ))[(offx + x + (32 as libc::c_int >> subx) * bx) as usize];
+            (offy + y + (32 >> suby) * by) as isize,
+        ))[(offx + x + (32 >> subx) * bx) as usize];
 }
 unsafe extern "C" fn fgy_32x32xn_c(
     dst_row: *mut pixel,
@@ -1012,7 +1012,7 @@ unsafe extern "C" fn fguv_32x32xn_c(
     let mut bx: libc::c_uint = 0 as libc::c_int as libc::c_uint;
     while (bx as size_t) < pw {
         let bw: libc::c_int = imin(
-            32 as libc::c_int >> sx,
+            32 >> sx,
             pw.wrapping_sub(bx as size_t) as libc::c_int,
         );
         if (*data).overlap_flag != 0 && bx != 0 {
@@ -1035,12 +1035,12 @@ unsafe extern "C" fn fguv_32x32xn_c(
             i_1 += 1;
         }
         let ystart: libc::c_int = if (*data).overlap_flag != 0 && row_num != 0 {
-            imin(2 as libc::c_int >> sy, bh)
+            imin(2 >> sy, bh)
         } else {
             0 as libc::c_int
         };
         let xstart: libc::c_int = if (*data).overlap_flag != 0 && bx != 0 {
-            imin(2 as libc::c_int >> sx, bw)
+            imin(2 >> sx, bw)
         } else {
             0 as libc::c_int
         };
@@ -1346,7 +1346,7 @@ unsafe extern "C" fn fguv_32x32xn_c(
             }
             y_0 += 1;
         }
-        bx = bx.wrapping_add((32 as libc::c_int >> sx) as libc::c_uint);
+        bx = bx.wrapping_add((32 >> sx) as libc::c_uint);
     }
 }
 unsafe extern "C" fn fguv_32x32xn_420_c(
@@ -1694,7 +1694,7 @@ unsafe extern "C" fn fguv_32x32xn_420_neon(
             type_0 as ptrdiff_t,
             bitdepth_max,
         );
-        bx = bx.wrapping_add((32 as libc::c_int >> 1) as libc::c_uint);
+        bx = bx.wrapping_add((32 >> 1) as libc::c_uint);
     }
 }
 
@@ -1780,7 +1780,7 @@ unsafe extern "C" fn fguv_32x32xn_422_neon(
             type_0 as ptrdiff_t,
             bitdepth_max,
         );
-        bx = bx.wrapping_add((32 as libc::c_int >> 1) as libc::c_uint);
+        bx = bx.wrapping_add((32 >> 1) as libc::c_uint);
     }
 }
 
@@ -1866,7 +1866,7 @@ unsafe extern "C" fn fguv_32x32xn_444_neon(
             type_0 as ptrdiff_t,
             bitdepth_max,
         );
-        bx = bx.wrapping_add((32 as libc::c_int >> 0) as libc::c_uint);
+        bx = bx.wrapping_add((32 >> 0) as libc::c_uint);
     }
 }
